Split to smaller components
