+++ src/components/weather/ObservationPanel.tsx
@@ -21,16 +21,16 @@ import { State } from '@store/types';
 import { useTheme } from '@react-navigation/native';
 import { useTranslation } from 'react-i18next';
 import { CustomTheme } from '@utils/colors';
-import moment from 'moment';
 import { TouchableOpacity } from 'react-native-gesture-handler';
-import Icon from '@components/common/Icon';
 import CloseButton from '@components/common/CloseButton';
 import Chart from './charts/Chart';
 import { ChartType } from './charts/types';
 import ParameterSelector from './charts/ParameterSelector';
 import CollapsibleHeader from './common/CollapsibleHeader';
 import PanelHeader from './common/PanelHeader';
-import WeatherInfoBottomSheet from './sheets/WeatherInfoBottomSheet';
+
+import List from './observation/List';
+import Latest from './observation/Latest';
 
 const mapStateToProps = (state: State) => ({
   data: selectData(state),
@@ -71,8 +71,6 @@ const ObservationPanel: React.FC<ObservationPanelProps> = ({
   const { colors } = useTheme() as CustomTheme;
   const { t } = useTranslation('observation');
   const infoSheetRef = useRef() as React.MutableRefObject<RBSheet>;
-  const weatherInfoSheetRef = useRef() as React.MutableRefObject<RBSheet>;
-  //  const [parameter, setParameter] = useState<ChartType>('temperature');
 
   useEffect(() => {
     const sid = stationList[0]?.id;
@@ -123,111 +121,7 @@ const ObservationPanel: React.FC<ObservationPanelProps> = ({
             />
           </View>
 
-          {data && data.length > 0 && (
-            <View>
-              <View style={[styles.row]}>
-                <View style={[styles.latestObservation]}>
-                  <Text style={[{ color: colors.shadow }]}>
-                    {t('latestObservation')}
-                  </Text>
-                  <Text
-                    style={[
-                      styles.bold,
-                      styles.justifyStart,
-                      { color: colors.shadow },
-                    ]}>
-                    {moment(data[data.length - 1].epochtime * 1000).format(
-                      `dd D.M. [${t('at')}] HH:mm`
-                    )}
-                  </Text>
-                </View>
-                <TouchableOpacity
-                  style={styles.bottomSheetButton}
-                  onPress={() => weatherInfoSheetRef.current.open()}>
-                  <Icon
-                    name="info"
-                    color={colors.shadow}
-                    height={24}
-                    width={24}
-                  />
-                </TouchableOpacity>
-              </View>
-              <View style={styles.observationRow}>
-                <View style={styles.observationPadding}>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.temperature')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.dewPoint')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.precipitation1h')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.windSpeedMS')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.windDirection')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.windGust')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.pressure')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.humidity')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.visibility')}
-                  </Text>
-                  <Text style={[styles.bold, { color: colors.shadow }]}>
-                    {t('measurements.totalCloudCover')}
-                  </Text>
-                </View>
-
-                <View>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {data[data.length - 1].temperature?.toFixed(1)} °C
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {data[data.length - 1].dewpoint?.toFixed(1)} °C
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {data[data.length - 1].precipitation1h != null
-                      ? data[data.length - 1].precipitation1h
-                      : '0.0'}{' '}
-                    mm
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {data[data.length - 1].windspeedms} m/s
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {t(`winddirection.${data[data.length - 1].windcompass8}`)} (
-                    {data[0].winddirection}
-                    °)
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {data[data.length - 1].windgust} m/s
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {data[data.length - 1].pressure} hPa
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {data[data.length - 1].humidity} %
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {Math.round(data[data.length - 1].visibility! / 1000)} km
-                  </Text>
-                  <Text style={[styles.panelText, { color: colors.shadow }]}>
-                    {t(`cloudcover.${data[data.length - 1].totalcloudcover}`)} (
-                    {data[data.length - 1].totalcloudcover}
-                    /8)
-                  </Text>
-                </View>
-              </View>
-            </View>
-          )}
+          <Latest data={data} />
         </View>
 
         <View style={styles.panelContainer}>
@@ -321,194 +215,7 @@ const ObservationPanel: React.FC<ObservationPanelProps> = ({
             parameter={parameter}
             setParameter={updateChartParameter}
           />
-          {toDisplay === LIST && (
-            <View style={styles.observationListContainer}>
-              <View>
-                {data && data.length > 0 && (
-                  <Text style={(styles.bold, styles.marginBottom)}>
-                    {moment(data[0].epochtime * 1000).format(`dddd D.M.`)}
-                  </Text>
-                )}
-              </View>
-              {parameter === 'temperature' && (
-                <View style={styles.observationRow}>
-                  <Text style={[styles.bold, styles.observationRow]}>Aika</Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Lämpötila
-                  </Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Kastepiste
-                  </Text>
-                </View>
-              )}
-              {parameter === 'precipitation' && (
-                <View style={styles.observationRow}>
-                  <Text style={[styles.bold, styles.observationRow]}>Aika</Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Sademäärä
-                  </Text>
-                </View>
-              )}
-              {parameter === 'wind' && (
-                <View style={styles.observationRow}>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    {'\n'}Aika
-                  </Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Tuulen{'\n'}nopeus
-                  </Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Tuulen{'\n'}puuska
-                  </Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Tuulen{'\n'}suunta
-                  </Text>
-                </View>
-              )}
-              {parameter === 'pressure' && (
-                <View style={styles.observationRow}>
-                  <Text style={[styles.bold, styles.observationRow]}>Aika</Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Paine
-                  </Text>
-                </View>
-              )}
-              {parameter === 'humidity' && (
-                <View style={styles.observationRow}>
-                  <Text style={[styles.bold, styles.observationRow]}>Aika</Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Kosteus
-                  </Text>
-                </View>
-              )}
-              {parameter === 'visCloud' && (
-                <View style={styles.observationRow}>
-                  <Text style={[styles.bold, styles.observationRow]}>Aika</Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Näkyvyys
-                  </Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Pilvisyys
-                  </Text>
-                </View>
-              )}
-              {parameter === 'cloud' && (
-                <View style={styles.observationRow}>
-                  <Text style={[styles.bold, styles.observationRow]}>Aika</Text>
-                  <Text style={[styles.bold, styles.observationRow]}>
-                    Pilven alarajan korkeus
-                  </Text>
-                </View>
-              )}
-
-              <View
-                style={
-                  (styles.marginBottom,
-                  {
-                    borderBottomWidth: 1,
-                    borderBottomColor: 'lightgray',
-                  })
-                }
-              />
-
-              {data
-                .filter((ob) => ob.epochtime % 3600 === 0)
-                .map((timeStep) => (
-                  <View key={timeStep.epochtime} style={styles.observationRow}>
-                    {parameter === 'temperature' && (
-                      <View style={styles.observationRow}>
-                        <Text style={[styles.bold, styles.observationRow]}>
-                          klo{' '}
-                          {moment(timeStep.epochtime * 1000).format(`HH:mm`)}
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {timeStep.temperature} °C
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {timeStep.dewpoint} °C
-                        </Text>
-                      </View>
-                    )}
-                    {parameter === 'precipitation' && (
-                      <View style={styles.observationRow}>
-                        <Text style={[styles.bold, styles.observationRow]}>
-                          klo{' '}
-                          {moment(timeStep.epochtime * 1000).format(`HH:mm`)}
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {timeStep.precipitation1h} mm
-                        </Text>
-                      </View>
-                    )}
-                    {parameter === 'wind' && (
-                      <View style={[styles.observationRow]}>
-                        <Text style={[styles.bold, styles.observationRow]}>
-                          klo{' '}
-                          {moment(timeStep.epochtime * 1000).format(`HH:mm`)}
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {timeStep.windspeedms?.toFixed(0)} m/s
-                        </Text>
-
-                        <Text style={styles.observationRow}>
-                          {timeStep.windgust?.toFixed(0)} m/s
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {timeStep.windcompass8}
-                        </Text>
-                      </View>
-                    )}
-                    {parameter === 'pressure' && (
-                      <View style={styles.observationRow}>
-                        <Text style={[styles.bold, styles.observationRow]}>
-                          klo{' '}
-                          {moment(timeStep.epochtime * 1000).format(`HH:mm`)}
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {timeStep.pressure?.toFixed(0)} hPa
-                        </Text>
-                      </View>
-                    )}
-                    {parameter === 'humidity' && (
-                      <View style={styles.observationRow}>
-                        <Text style={[styles.bold, styles.observationRow]}>
-                          klo{' '}
-                          {moment(timeStep.epochtime * 1000).format(`HH:mm`)}
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {timeStep.humidity?.toFixed(0)} %
-                        </Text>
-                      </View>
-                    )}
-                    {parameter === 'visCloud' && (
-                      <View style={[styles.observationRow]}>
-                        <Text style={[styles.bold, styles.observationRow]}>
-                          klo{' '}
-                          {moment(timeStep.epochtime * 1000).format(`HH:mm`)}
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {(timeStep.visibility! / 1000)?.toFixed(0)} km
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {timeStep.totalcloudcover}/8
-                        </Text>
-                      </View>
-                    )}
-                    {parameter === 'cloud' && (
-                      <View style={styles.observationRow}>
-                        <Text style={[styles.bold, styles.observationRow]}>
-                          klo{' '}
-                          {moment(timeStep.epochtime * 1000).format(`HH:mm`)}
-                        </Text>
-                        <Text style={styles.observationRow}>
-                          {(timeStep.cloudheight! / 1000)?.toFixed(1)} km
-                        </Text>
-                      </View>
-                    )}
-                  </View>
-                ))}
-            </View>
-          )}
+          {toDisplay === LIST && <List data={data} parameter={parameter} />}
           {toDisplay === CHART && (
             <Chart chartType={parameter} data={data} observation />
           )}
@@ -556,20 +263,6 @@ const ObservationPanel: React.FC<ObservationPanelProps> = ({
           ))}
         </View>
       </RBSheet>
-      <RBSheet
-        ref={weatherInfoSheetRef}
-        height={600}
-        closeOnDragDown
-        customStyles={{
-          container: {
-            ...styles.sheetContainer,
-            backgroundColor: colors.background,
-          },
-        }}>
-        <WeatherInfoBottomSheet
-          onClose={() => weatherInfoSheetRef.current.close()}
-        />
-      </RBSheet>
     </View>
   );
 };
@@ -578,9 +271,6 @@ const styles = StyleSheet.create({
   bold: {
     fontFamily: 'Roboto-Bold',
     paddingBottom: 2,
-  },
-  bottomSheetButton: {
-    padding: 10,
   },
   bottomSheetTitle: {
     marginLeft: 15,
@@ -608,30 +298,11 @@ const styles = StyleSheet.create({
   justifyStart: {
     justifyContent: 'flex-start',
   },
-  latestObservation: {
-    flexDirection: 'column',
-    paddingBottom: 5,
-  },
-  marginBottom: {
-    marginBottom: 20,
-  },
   medium: {
     fontFamily: 'Roboto-Medium',
   },
   observationDropdown: {
     marginBottom: 10,
-  },
-  observationListContainer: {
-    marginTop: 20,
-  },
-  observationPadding: {
-    paddingRight: 100,
-  },
-  observationRow: {
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    flex: 1,
-    paddingBottom: 1,
   },
   observationText: {
     fontSize: 14,
